# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: number;
  amount: number;
  category: string;
  type: TransactionType;
  date: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  type: TransactionInputType;
  date: string;
  note?: string;
}

export interface FinancialSummary {
  totalIncome: number;
  totalExpense: number;
  netBalance: number;
  /** @nullable */
  topSpendingCategory: string | null;
  transactionCount: number;
}

export interface CategorySpending {
  category: string;
  total: number;
  count: number;
}

export type SpendingInsightSeverity = typeof SpendingInsightSeverity[keyof typeof SpendingInsightSeverity];


export const SpendingInsightSeverity = {
  info: 'info',
  warning: 'warning',
  positive: 'positive',
} as const;

export interface SpendingInsight {
  type: string;
  message: string;
  severity: SpendingInsightSeverity;
}

export interface ErrorResponse {
  error: string;
}

export type ListTransactionsParams = {
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by type
 */
type?: ListTransactionsType;
/**
 * Filter from date (YYYY-MM-DD)
 */
dateFrom?: string;
/**
 * Filter to date (YYYY-MM-DD)
 */
dateTo?: string;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
} as const;