# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: summary
    description: Summary and analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      description: List all transactions with optional filters
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income, expense]
          description: Filter by type
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter from date (YYYY-MM-DD)
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter to date (YYYY-MM-DD)
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /summary:
    get:
      operationId: getSummary
      tags: [summary]
      summary: Get financial summary
      description: Returns total income, total expense, net balance, and top spending category
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

  /summary/by-category:
    get:
      operationId: getSpendingByCategory
      tags: [summary]
      summary: Get spending by category
      description: Returns aggregated expense totals grouped by category
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"

  /summary/insight:
    get:
      operationId: getInsight
      tags: [summary]
      summary: Get spending insight
      description: Returns a rule-based insight derived from spending patterns
      responses:
        "200":
          description: Spending insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendingInsight"

  /categories:
    get:
      operationId: listCategories
      tags: [transactions]
      summary: List all categories
      description: Returns all unique categories used in transactions
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      required: [id, amount, category, type, date, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
          format: float
        category:
          type: string
        type:
          type: string
          enum: [income, expense]
        date:
          type: string
          format: date
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    TransactionInput:
      type: object
      required: [amount, category, type, date]
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
        category:
          type: string
          minLength: 1
        type:
          type: string
          enum: [income, expense]
        date:
          type: string
          format: date
        note:
          type: string

    FinancialSummary:
      type: object
      required: [totalIncome, totalExpense, netBalance, topSpendingCategory, transactionCount]
      properties:
        totalIncome:
          type: number
          format: float
        totalExpense:
          type: number
          format: float
        netBalance:
          type: number
          format: float
        topSpendingCategory:
          type: ["string", "null"]
        transactionCount:
          type: integer

    CategorySpending:
      type: object
      required: [category, total, count]
      properties:
        category:
          type: string
        total:
          type: number
          format: float
        count:
          type: integer

    SpendingInsight:
      type: object
      required: [type, message, severity]
      properties:
        type:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, positive]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string