# lib/api-zod/src/generated/types/categorySpending.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySpending {
  category: string;
  total: number;
  count: number;
}