# lib/api-zod/src/generated/types/financialSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialSummary {
  totalIncome: number;
  totalExpense: number;
  netBalance: number;
  /** @nullable */
  topSpendingCategory: string | null;
  transactionCount: number;
}