# lib/api-zod/src/generated/types/spendingInsight.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SpendingInsightSeverity } from './spendingInsightSeverity';

export interface SpendingInsight {
  type: string;
  message: string;
  severity: SpendingInsightSeverity;
}