# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from './transactionType';

export interface Transaction {
  id: number;
  amount: number;
  category: string;
  type: TransactionType;
  date: Date;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}