# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputType } from './transactionInputType';

export interface TransactionInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  type: TransactionInputType;
  date: Date;
  note?: string;
}